backend/pyproject.toml
[project]
name = "memomap-learning-backend"
version = "0.1.0"
description = "Memomap Learning Backend API"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "alembic>=1.14.0",
    "azure-cognitiveservices-speech>=1.41.1",
    "flask>=3.1.2",
    "flask-cors>=6.0.1",
    "flask-limiter>=3.13",
    "google-auth>=2.0.0",
    "google-genai>=1.49.0",
    "itsdangerous>=2.2.0",
    "loguru>=0.7.3",
    "prometheus-client>=0.23.0",
    "psycopg2-binary>=2.9.0",
    "pycountry>=24.6.1",
    "pydantic>=2.11.9",
    "pyjwt>=2.8.0",
    "python-dotenv>=1.1.1",
    "redis>=7.0.1",
    "requests>=2.32.5",
    "sqlalchemy>=2.0.0",
]

[project.scripts]
dev = "run:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src", "run.py"]
